remove other themes
