--- packages/react-components/theme-designer/src/components/Sidebar/EditTab.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-export interface EditTabProps {}
-
-export const EditTab: React.FC<EditTabProps> = props => {
-  return <div role="tabpanel" aria-labelledby="Edit" />;
-};
